requirements.txt
torch==1.9.0
torchvision==0.10.0
matplotlib
opencv-python
lpips
tqdm
imageio
pytorch_fid
pytorch-msssim
argparse
pillow
numpy
pytorch_msssim